Add files via upload
Badger's Law and my goofy ass is proposing that the Yin-Yang symbol exits as potentially a sphere and the two dots are always touching and never connect. And they're connected by an Einstein-Rosen-TommyLongsleeves Spiral Bridge. Any two points on the sphere are the same distance thanks to our spirals!!! Woo! - This has to be the best gayness science ever! Woo! - YEs we want to test it more too - no sh*t.